Bugfix: do not deep merge the old and new inputs
For primitive values like ints, this doesn't matter, but for things like
attrsets (and derivations are attrsets), you do not want to merge the
old values and the new values, we *only* want the new values.

I ran into this bug with a flake's output derivation, which started
acting quite strange because the old and new derivation were getting
merged.